feat: updates

+++ Bank_System/Currency/Currency.cs
@@ -1,3 +1,3 @@
 namespace Bank_System;
 
-public readonly record struct Currency(string Cc, double Rate);
+public readonly record struct Currency(string Cc, decimal Rate);
